Create twitter_scrape_n_count_with_trigger.json
This NodeRed flow searches Twitter for certain hashtags (in this case water and fail), pushes the Tweet to an output and counts the number of Tweets retrieved. After 5 Tweets, it triggers a signal and resets the counter. It's part of a project that will grow, which will interact with an RPi to control a valve to release water on an accumulation of Tweets.